requirements.txt
aiogram==3.11.0
aiogram_dialog==2.2.0
python-dotenv==1.0.1
redis==5.1.0